requirements.txt
# Autonomous Research Swarm - Dependencies

# Core async and HTTP
httpx>=0.27.0
aiohttp>=3.9.0
aiofiles>=23.2.0

# Web scraping and content extraction
beautifulsoup4>=4.12.0
trafilatura>=1.9.0
requests>=2.31.0

# Text processing and NLP
langdetect>=1.0.9
python-levenshtein>=0.25.0
fuzzy-wuzzy>=0.18.0

# URL and domain handling
urlextract>=1.9.0
tldextract>=5.1.0

# Data structures and validation
pydantic>=2.7.0
dataclasses-json>=0.6.0

# Rate limiting and retries
ratelimit>=2.2.1
backoff>=2.2.1
tenacity>=8.2.0

# Date and time handling
python-dateutil>=2.9.0

# Configuration and environment
PyYAML>=6.0.1
python-dotenv>=1.0.0

# Vector databases and embeddings (optional)
chromadb>=0.5.0
faiss-cpu>=1.8.0; platform_system != "Windows"
numpy>=1.24.0
scipy>=1.11.0

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0

# CLI and utilities
click>=8.1.0
rich>=13.7.0
tqdm>=4.66.0

# Logging and monitoring
loguru>=0.7.0

# Security and safety
cryptography>=41.0.0

# Optional: Jupyter for analysis
jupyter>=1.0.0
ipykernel>=6.25.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.1.0

# Optional: Advanced NLP
transformers>=4.35.0
torch>=2.1.0; platform_system != "Darwin"
sentence-transformers>=2.2.0

# Optional: Database support
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Optional: API integrations
openai>=1.3.0
anthropics>=0.8.0
cohere>=4.37.0

# Development tools
black>=23.9.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.6.0
pre-commit>=3.5.0